requirements.txt
fastapi
uvicorn
transformers
torch
pillow
sqlalchemy
psycopg2-binary
python-multipart